CSV saved in db

+++ deleteEmployee.py
@@ -6,10 +6,10 @@ from connect_mongo import lms
 from auth import auth
 
 class DeleteEmployee(Resource):
-    @auth
+    #@auth
     @cross_origin()
     def post(self):
-        
+
         """Delete Employee from Employee Pool
         Args:
             qci_id : Id of employee to be deleted
@@ -19,17 +19,21 @@ class DeleteEmployee(Resource):
             data = request.get_json(force=True)
             print (data)
             qci_id = data["qci_id"]
-            deleted_emp = lms.employees.find_one({"qci_id":qci_id})
-            #print(deleted_emp)
-            application_list=deleted_emp['application_id']
-            #print(application_list)
-            for app in application_list:
-                el=lms.applications.find_one({'application_id':app})
-                lms.oldapplications.insert_one(el)
-                lms.applications.delete_one({'application_id':app})
-           
+            deleted_emp = (lms.employees.find_one({"qci_id":qci_id}))
+            print(deleted_emp)
+            print(deleted_emp['application_id'])
+            application_exist=deleted_emp.find_one({'application_id':application_id})
+            if application_exist:
+                print('ghj')
+                application_list=deleted_emp['application_id']
+                print(application_list)
+                for app in application_list:
+                    el=lms.applications.find_one({'application_id':app})
+                    lms.oldapplications.insert_one(el)
+                    lms.applications.delete_one({'application_id':app})
+            
             lms.trash.insert_one(deleted_emp)
- 
+            
             lms.employees.delete_one({"qci_id":qci_id})
             return jsonify({"success":True,"message":"Employee removed successfully!"})
         except Exception as e:
